chore: improved initialize and textdoc/completion methods
